sync and async version of download
