+++ get_euler_solved_count.py
@@ -1,33 +1,17 @@
 import time
 import json
-import asyncio
-import aiohttp
 import requests
 from bs4 import BeautifulSoup
 
-urls = ['http://projecteuler.net/problem={}'.format(i)
-        for i in range(1, 691)]
-
-async def async_get(session, url):
-    r = await session.request('GET', url=url, timeout=60)
-    d = await r.content.read()
-    return url, d
-
-async def async_api_request(urls=urls):
-    semaphore = asyncio.Semaphore(10)
-    async with semaphore:
-        async with aiohttp.ClientSession() as session:
-            tasks = []
-            for url in urls:
-                tasks.append(async_get(session=session, url=url))
-            results = await asyncio.gather(*tasks)
-            return results
 
 def get_problem_number(url):
     return url.split('=')[-1]
 
-def handle_response(content):
-    content = content.decode()
+
+def get_problem_info(problem_number):
+    url = f'http://projecteuler.net/problem={problem_number}'
+    r = requests.get(url, timeout=1)
+    content = r.content.decode()
     soup = BeautifulSoup(content, 'html.parser')
     section = str(soup.find_all('div')[0]).split(';')
     for text in section:
@@ -35,21 +19,17 @@ def handle_response(content):
             return int(text.split(' ')[-1])
     return
 
-def handle_results(results):
-    return {get_problem_number(url): handle_response(r)
-            for url, r in results}
 
-def download():
+def download_results():
     start = time.time()
-    results = async_api_request(urls)
-    loop = asyncio.get_event_loop()
-    data = loop.run_until_complete(results)
-    records = handle_results(data)
+    results = {problem_number: get_problem_info(problem_number)
+               for problem_number in range(1, 691)}
     with open("project_euler_solve_counts.json", "w") as f:
-        json.dump(records, f)
+        json.dump(results, f)
     end = time.time()
     print(f"Time taken: {round(end-start, 3)}s")
     return
 
+
 if __name__ == "__main__":
-    download()
+    download_results()
